update half float and seq no
